pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openagno"
version = "1.1.0"
description = "Build autonomous AI agents with declarative YAML configuration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "agno[os,scheduler]>=1.5.0",
  "aioboto3>=13.0",
  "anthropic>=0.30.0",
  "arxiv>=2.0.0",
  "boto3>=1.35",
  "crawl4ai>=0.4.0",
  "chardet>=5.2,<6",
  "ddgs>=6.0",
  "duckduckgo-search>=6.0",
  "fastapi>=0.110.0",
  "fastmcp>=2.0.0",
  "google-genai>=1.0.0",
  "httpx>=0.27.0",
  "mcp>=1.0.0",
  "mistralai>=1.0.0",
  "openai>=1.30.0",
  "pgvector>=0.3",
  "psycopg[binary]>=3.1",
  "PyGithub>=2.0",
  "python-dotenv>=1.0.0",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0",
  "rich>=13.0",
  "slowapi>=0.1.9",
  "sqlalchemy>=2.0",
  "tavily-python>=0.4.0",
  "typer>=0.12.0",
  "uvicorn>=0.27.0",
  "wikipedia>=1.4.0",
  "wsproto>=1.3.0",
  "yfinance>=0.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.0.0",
]

[project.scripts]
openagno = "openagno.cli:main"

[tool.setuptools]
include-package-data = true
py-modules = ["gateway", "loader", "security", "service_manager"]

[tool.setuptools.packages.find]
include = ["management*", "openagno*", "routes*", "tools*"]

[tool.setuptools.package-data]
openagno = ["templates/**/*.md", "templates/**/*.yaml"]
